Move BooleanConverters to its own file
